mysql

+++ MMSEngine/src/MMSEngineDBFacade_EncodingJobs.cpp
@@ -4659,7 +4659,7 @@ Json::Value MMSEngineDBFacade::getEncodingJobsStatus (
                 if (resultSetEncodingJob->isNull("encodingProgress"))
                     encodingJobRoot[field] = Json::nullValue;
                 else
-                    encodingJobRoot[field] = resultSetEncodingJob->getFloat("encodingProgress");
+                    encodingJobRoot[field] = (double) resultSetEncodingJob->getDouble("encodingProgress");
 
                 field = "start";
                 if (encodingStatus == EncodingStatus::ToBeProcessed)
